+++ CHANGELOG.md
@@ -1,4 +1,17 @@
 
+## Unreleased
+---
+
+### New
+
+### Changes
+
+### Fixes
+
+### Breaks
+
+
+
 
 # CHANGELOG
 
@@ -8,10 +21,9 @@ Changelog](http://keepachangelog.com/).
 
 ## master
 
-### Fixes
 * Fix error with ARGS
 
-### New
+* Multi-task version.
 
 * Experimental feature of async with async parameter in server. It is nicely working but it can merge the outputs in console.